Create TOwerof Hanoi.c
 Tower of Hanoi
 follows these rules:

1. Only one disk can be moved at a time.
2. A larger disk cannot be placed on top of a smaller disk.
3. All disks start on one peg, and the goal is to move them to another peg using an auxiliary peg.

Code Breakdown:
#include <stdio.h>:

This is the standard input-output library in C. It is included to allow the use of printf to display the output.
Function to:

Void toh(int n, char a, char b, char c): This function solves the Tower of Hanoi problem recursively.

n: The number of disks to move.
a: The source peg.
b: The destination peg.
c: The auxiliary peg.
Base Case (if (n == 1)):

If there is only one disk, it simply moves it from the source peg a to the destination peg b. It prints the move as a -> b.
Recursive Case:

Move the top n-1 disks from peg a (source) to peg c (auxiliary), using b (destination) as a temporary peg.
Then, move the largest disk (the nth disk) from peg a (source) to peg b (destination).
Finally, move the n-1 disks from peg c (auxiliary) to peg b (destination), using an as a temporary peg.
Main Function (int main()):

int n = 3; The number of disks is set to 3.
toh(n, 'a', 'b', 'c');: The toh function is called with the following arguments:
n = 3: Move 3 disks.
'a': The source peg.
'b': The destination peg.
'c': The auxiliary peg.
Output: The program will print the steps to solve the Tower of Hanoi puzzle for 3 disks. The expected output is a series of moves showing how to transfer the disks from Peg A to peg b using Peg C.